Split the channels 2 app into different modules
